+++ lib/ModelMixins/GeojsonMixin.ts
@@ -76,6 +76,7 @@ import Model, { BaseModel } from "../Models/Definition/Model";
 import StratumOrder from "../Models/Definition/StratumOrder";
 import TableAutomaticStylesStratum from "../Table/TableAutomaticStylesStratum";
 import { GeoJsonTraits } from "../Traits/TraitsClasses/GeoJsonTraits";
+import LegendTraits from "../Traits/TraitsClasses/LegendTraits";
 import { RectangleTraits } from "../Traits/TraitsClasses/MappableTraits";
 import { DiscreteTimeAsJS } from "./DiscretelyTimeVaryingMixin";
 import { ExportData } from "./ExportableMixin";
@@ -299,6 +300,22 @@ function GeoJsonMixin<T extends Constructor<Model<GeoJsonTraits>>>(Base: T) {
     @computed
     get disableSplitter() {
       return !this._imageryProvider;
+    }
+
+    /** Special case for legends.
+     * Because TableMixin does not have LegendTraits, but GeoJsonMixin does, we need to check to see if traits have been defined.
+     * If so, they will override TableMixin legends
+     */
+    @computed
+    get legends(): Model<LegendTraits>[] {
+      const legendTraits = this.traits.legends.getValue(this) as
+        | Model<LegendTraits>
+        | undefined;
+      if (Array.isArray(legendTraits) && legendTraits.length > 0) {
+        return legendTraits;
+      }
+
+      return super.legends;
     }
 
     @computed get mapItems() {