+++ lib/Traits/TraitsClasses/ArcGisTerrainCatalogItemTraits.ts
@@ -1,10 +1,12 @@
-import CatalogMemberTraits from "./CatalogMemberTraits";
-import MappableTraits from "./MappableTraits";
 import mixTraits from "../mixTraits";
+import CatalogMemberTraits from "./CatalogMemberTraits";
+import LegendOwnerTraits from "./LegendOwnerTraits";
+import MappableTraits from "./MappableTraits";
 import UrlTraits from "./UrlTraits";
 
 export default class ArcGisTerrainCatalogItemTraits extends mixTraits(
   UrlTraits,
   MappableTraits,
-  CatalogMemberTraits
+  CatalogMemberTraits,
+  LegendOwnerTraits
 ) {}