+++ lib/Traits/TraitsClasses/WebProcessingServiceCatalogGroupTraits.ts
@@ -5,12 +5,14 @@ import GetCapabilitiesTraits from "./GetCapabilitiesTraits";
 import GroupTraits from "./GroupTraits";
 import mixTraits from "../mixTraits";
 import UrlTraits from "./UrlTraits";
+import LegendOwnerTraits from "./LegendOwnerTraits";
 
 export default class WebProcessingServiceCatalogGroupTraits extends mixTraits(
   UrlTraits,
   GroupTraits,
   GetCapabilitiesTraits,
-  CatalogMemberTraits
+  CatalogMemberTraits,
+  LegendOwnerTraits
 ) {
   @anyTrait({
     name: "Item Properties",