+++ CHANGES.md
@@ -12,6 +12,9 @@ Change Log
 * Fix `SplitReferences` which use `shareKeys`
 * Make `isJson*` type assertion functions more rigorous
   * Add `deep` parameter, so you can use old "shallow" type check for performance reasons if needed
+* Add Shapefile to `CkanDefaultFormatsStratum`
+* Fix `ArcGisMapServerCatalogItem` metadata bug
+* Remove legend traits from CatalogMemberMixin, replacing them with LegendOwnerTraits, and add tests to enforce correct use of legends.
 * Add better support for retreiving GeoJsonCatalogItem data through APIs, including supporting geojson nested within json objects
 * Fixed `ContinuousColorMap` min/max value bug.
 * `TableStyle.outlierColor` is now only used if `zFilter` is active, or `colorTraits.outlierColor` is defined
@@ -21,6 +24,8 @@ Change Log
 * Add `WebMapServiceCatalogItem.rectangle` support for multiple WMS layers
 * Fix picked feature highlighting for ArcGis REST API features (and TSify `featureDataToGeoJson`)
 * Re-enable GeoJSON simple styling - now if more than 50% of features have [simple-style-spec properties](https://github.com/mapbox/simplestyle-spec) - automatic styling will be disabled (this behaviour can be disabled by setting `forceCesiumPrimitives = false`)
+* Don't show `TableMixin` `legends` or `mapItems` if no data
+* Fix `GeoJsonCatalogItem.legends`
 * [The next improvement]
 
 #### 8.1.14